requirements.txt
Flask==2.1.2
Flask-SQLAlchemy==2.5.1
Flask-JWT-Extended==4.3.1
python-dotenv==0.19.2
SQLAlchemy==1.3.23
werkzeug==2.2.2
Flask-CORS==3.0.10  # Tambahkan Flask-CORS
